+++ benches/ezsockets_server.rs
@@ -52,7 +52,7 @@ impl ezsockets::SessionExt for EchoSession {
     }
 
     async fn on_text(&mut self, text: String) -> Result<(), ezsockets::Error> {
-        self.handle.text(text);
+        self.handle.text(text).unwrap();
         Ok(())
     }